user-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
psycopg2-binary==2.9.7